refactor(requests): drop dead StatusBadge, relocate RequestStatus type
StatusBadge was never rendered (only its type was imported) and its
palette contradicted the canonical StatusPill. Delete the component,
move the RequestStatus type to the data layer (lib/supabase/requests),
and repoint the two type importers. Also swaps RequestHistory's search
bg to the new sbi-input token.
