+++ frontend/components/dashboard/requests/index.tsx
@@ -2,4 +2,3 @@ export { FileUpload } from "./FileUpload";
 export { RequestDetailModal } from "./RequestDetailModal";
 export { RequestForm } from "./RequestForm";
 export { RequestHistory } from "./RequestHistory";
-export { StatusBadge } from "./StatusBadge";
